+++ opm/common/utility/numeric/SparseVector.hpp
