+++ opm/common/utility/numeric/linearInterpolation.hpp
